remover helpers and filters from details

+++ entities/Details.go
@@ -4,7 +4,7 @@ import "time"
 
 type Detail struct {
 	ID               int       `json:"id" db:"id_detail"`
-	PlantId          int       `json:"plantId" db:"id_plant"`
+	PlantId          int64     `json:"plantId" db:"id_plant"`
 	Time             time.Time `json:"time" db:"time"`
 	InternalHumidity float32   `json:"internalHumidity" db:"internal_humidity" binding:"required"`
 	ExternalHumidity float32   `json:"externalHumidity" db:"external_humidity" binding:"required"`
